Bug fix in printf output, removed barrier
There was a bug in the printf output.  The value should have be 'pe+me' not
'npes+pe'.  An unnecessary explicit barrier was also removed. The formatting of
the code was made consistent with other (4 to 3 spaces for a tab).  The global
sync array was made to static and a barrier is only needed to for
initialization of source and destination arrays.  The sync array is now
initialized before the implicit barier in shmem_init().