Bug fix in printf, now strides, removed barrier
There was a bug in the printf output.  The value should have be 'pe+me' not
'npes+pe'.  The code also did not demonstrate the strided all-to-all capability
and did not improve upon the non-strided example.  So strides are now used and
results are verified.

An unnecessary explicit barrier was also removed. The formatting of the code
was made consistent with other (4 to 3 spaces for a tab).  The global sync
array was made to static and a barrier is only needed to for initialization of
source and destination arrays.  The sync array is now initialized before the
implicit barier in shmem_init().