Bug fix for non-deterministic behavior
The previous code had non-deterministic behavior. The local value of x may be
set after the remote PE sets the local value.  Removing 'x = 1000 + me' brings
the code in line with barrierall example

The global sync array was made to static and is now initialized before the
implicit barier in shmem_init().