Added variable nelems for shmem_collect32()
The previous code used a constant number of elements per PE which is more of a
demonstration of the shmem_fcollect32() routine, not shmem_collect32(). This
patch now has a variable number of elements from each PE source with a
deterministic result.

The final output was also unreadable so a locking scheme is used.

The global sync array was made to static and is now initialized before the
implicit barrier in shmem_init().

Other formatting improvements are made