Removed barrier, global to static scope, C99 decl
The relevant variables have moved from global to static scope and use C99
declaration/initialization.  They are set before the implicit barrier in
shmem_init() so that the explicity shmem_barrier_all() barrier is no longer
necessary.  The shmem_malloc is removed for simplicity and a static variable is
used instead.