Consistency edits from pages 60 to 85.
This PR is the third part of a series of three. The purpose of the consis_edit PRs is to make the OpenSHMEM 1.4 specification, specifically the 8.1-8.10 sections, read more consistently. Edits have been made to section intros, routine intros, API descriptions, example sections, and argument lists with the intention of making them read as though they were written by one person. This PR includes consistency edits from pages 60 to 85.