Consistency edits for the term: active set.
The purpose of this file is to change the formatting of the term ‘active set,’ and one instance of the term ‘collective routines,’ so that they are consistent with the way other important terms appear in the spec. The following is an explanation justifying the edit:

When an important term is introduced in the spec it is written in italics and every first letter is capitalized. Symmetric data objects, for example, reads as Symmetric Data Objects when first introduced. Every subsequent time an important term appears, all letters are written in the lower case and the term is not italicized. The term ‘active set,’ when it first appears, reads like so: Active set. Collective routines, when it first appears, also follows this formatting. Almost every subsequent time ‘active set’ appears in the spec the term is written this way. The a is capitalized and the term itself is written in italics. No other important terms follow this pattern.