section/collectives: reorder reduction op for active-set colls
The reduction operators mentioned in the intro of active-set-based
collectives did not follow a specific order, i.e., neither alphabetized
nor grouped by operator types. This patch reorders it to match the
description of team-based collectives. The new order also matches that
of the operator definitions in section 9.11.7 SHMEM_REDUCTIONS.