Reupdate full/partial completion to signal_wait_until
We were addressing the partial update completions in the apinotes
for the shmem_wait_until_signal op. To be consistent with other
sync operations - moving to the description section.

This commit was previously pushed in the openshmem-org:master and
it had a hash clash and the change wasn;t pushed in the master
branch. We are fixing it.